ai_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai==1.51.0
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
